Refactor: move code into lifecycle.package
lifecycle.assemble was an older incarnation of lifecycle.package, so
moving tests out. And lifecycle.distribute.bytes wasn't really doing
anything bytes specific. (At least not enough to warrant being separated
from the other packaging code)

No changes to code in the files. Just moving things around.
